% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keys.R
\name{fk_ish}
\alias{fk_ish}
\title{Check a tbl for a candidate foreign key constraint relationship}
\usage{
fk_ish(x, y, by = NA_character_,
  na_matches = pkgconfig::get_config("dplyr::na_matches"))
}
\arguments{
\item{x}{data frame}

\item{y}{data frame with foreign key relationship to x}

\item{by}{character vector of key relationship, with semantics of dplyr join \code{by} parameters.}
}
\value{
boolean whether \code{by} variables could specify a foreign key relationship \code{x}->\code{y}

Satisfying a foreign key constraint relationship would ensure that
\code{z <- dplyr::left_join(x, y, by = by)} would result in all rows of \code{x} appear in \code{z}.
A foreign key constraint requires two things:
\enumerate{
\item that all the key values in x be present in y
\item that key be pk_ish in y (e.g., distinct combinations)
Note: currently using \code{by} parameter as a character vector, but join verbs have
more flexible specification. When this is addressed, probably the default arg will be NULL.
}
}
\description{
Check a tbl for a candidate foreign key constraint relationship
}
\examples{
x <- tibble(id = c(1, 2, 3))
y <- tibble(id = c(1, 2, 3), val = c(100, 200, 300))
by <- "id"
fk_ish(x, y, "id")
}
