% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{left_join_fk}
\alias{left_join_fk}
\title{Left-join while checking that \code{.y} has primary key \code{by}}
\usage{
left_join_fk(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...,
  na_matches = pkgconfig::get_config("dplyr::na_matches"))
}
\arguments{
\item{x}{a data frame}

\item{y}{a data frame}

\item{by}{a character vector specification of variables to join by (with a foreign key relationship)}
}
\description{
Wraps \code{dplyr::left_join()} with extra checking and diagnostics.
}
\examples{
if (require("Lahman")) {
batting_df <- tbl_df(Batting)
person_df <- tbl_df(Master)
uperson_df <- tbl_df(Master[!duplicated(Master$playerID), ])

left_join_pk(batting_df, uperson_df)
left_join_pk(batting_df, person_df)
}
}
