% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keys.R
\name{pk_ish}
\alias{pk_ish}
\title{Check a tbl for a candidate primary key}
\usage{
pk_ish(.data, by = NULL,
  na_matches = pkgconfig::get_config("dplyr::na_matches"))
}
\arguments{
\item{by}{a character vector specifying a candidate primary to (e.g., to check)}

\item{x}{a data frame}
}
\value{
boolean whether \code{x} has primary key \code{by}
A primary key of a table is a column (or set of columns) which is distinct.
}
\description{
Check a tbl for a candidate primary key
}
\examples{
dfx <- tibble::tibble(id = c(1, 2, 3), val = c(1, 1, 2))
pk_ish(dfx, "id")
dfy <- tibble::tibble(id = c(1, 1, 2), val = c(1, 2, 3))
pk_ish(dfy, "id")
}
