% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblrelations-package.R
\docType{package}
\name{tblrelations-package}
\alias{tblrelations}
\alias{tblrelations-package}
\title{tblrelations: Extra diagnostics for joins}
\description{
Basic tools based on dplyr when you use R like relational DB tables
}
\author{
\strong{Maintainer}: Jameel Alsalam \email{jalsalam@gmail.com}

}
\keyword{internal}
